+++ libs/btp/navigation-menu/src/index.ts
@@ -1,0 +1,5 @@
+export * from './lib/navigation-menu-item.component';
+export * from './lib/navigation-menu-popover-control.directive';
+export * from './lib/navigation-menu-popover.component';
+export * from './lib/navigation-menu.component';
+export * from './lib/navigation-menu.module';