+++ libs/btp/tool-header/src/tool-header-action-button.type.ts
@@ -1,0 +1,11 @@
+import { TemplateRef } from '@angular/core';
+
+export interface FdbToolHeaderActionButton {
+    templateRef?: TemplateRef<void>;
+    glyph: string;
+    label: string;
+    forceVisibility?: boolean;
+    clickCallback: () => void;
+    hasBadge?: boolean;
+    hasSeparator?: boolean;
+}