+++ libs/docs/btp/tool-header/tool-header-docs.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { FocusableGridDirective } from '@fundamental-ngx/cdk/utils';
+import { MessageStripComponent } from '@fundamental-ngx/core/message-strip';
 import { TableModule } from '@fundamental-ngx/core/table';
 import {
     CodeExampleComponent,
@@ -7,9 +8,15 @@ import {
     DescriptionComponent,
     DocsSectionTitleComponent,
     ExampleFile,
-    getExampleFile
+    getExampleFile,
+    SeparatorComponent
 } from '@fundamental-ngx/docs/shared';
+import { ToolHeaderAutoModeExampleComponent } from './examples/auto-mode-example/auto-mode-example.component';
 import { ToolHeaderBasicExampleComponent } from './examples/basic-example/tool-header-basic-example.component';
+import { CustomLogoExampleComponent } from './examples/custom-logo-example/custom-logo-example.component';
+import { MenuButtonExampleComponent } from './examples/menu-button-example/menu-button-example.component';
+import { ProductSwitchExampleComponent } from './examples/product-switch-example/product-switch-example.component';
+import { UsageWithLayoutExampleComponent } from './examples/usage-with-layout-example/usage-with-layout-example.component';
 
 @Component({
     templateUrl: './tool-header-docs.component.html',
@@ -21,14 +28,48 @@ import { ToolHeaderBasicExampleComponent } from './examples/basic-example/tool-h
         TableModule,
         ComponentExampleComponent,
         CodeExampleComponent,
-        ToolHeaderBasicExampleComponent
+        ToolHeaderBasicExampleComponent,
+        SeparatorComponent,
+        ToolHeaderAutoModeExampleComponent,
+        ProductSwitchExampleComponent,
+        MenuButtonExampleComponent,
+        CustomLogoExampleComponent,
+        MessageStripComponent,
+        UsageWithLayoutExampleComponent
     ]
 })
 export class ToolHeaderDocsComponent {
     basicExample: ExampleFile[] = [
+        getExampleFile('basic-example/tool-header-basic-example.component.html'),
         getExampleFile('basic-example/tool-header-basic-example.component.ts', {
             component: 'ToolHeaderBasicExampleComponent'
-        }),
-        getExampleFile('basic-example/tool-header-basic-example.component.ts')
+        })
+    ];
+    autoModeExample: ExampleFile[] = [
+        getExampleFile('auto-mode-example/auto-mode-example.component.html'),
+        getExampleFile('auto-mode-example/auto-mode-example.component.ts', {
+            component: 'ToolHeaderAutoModeExampleComponent'
+        })
+    ];
+    toolLayoutExample: ExampleFile[] = [
+        getExampleFile('usage-with-layout-example/usage-with-layout-example.component.html'),
+        getExampleFile('usage-with-layout-example/usage-with-layout-example.component.ts', {
+            component: 'UsageWithLayoutExampleComponent'
+        })
+    ];
+    menuButtonExample: ExampleFile[] = [
+        getExampleFile('menu-button-example/menu-button-example.component.ts', {
+            component: 'MenuButtonExampleComponent'
+        })
+    ];
+    productSwitchExample: ExampleFile[] = [
+        getExampleFile('product-switch-example/product-switch-example.component.ts', {
+            component: 'ProductSwitchExampleComponent'
+        })
+    ];
+    customLogoExample: ExampleFile[] = [
+        getExampleFile('custom-logo-example/custom-logo-example.component.ts', {
+            component: 'CustomLogoExampleComponent'
+        })
     ];
 }