+++ CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.46.0-rc.92](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.91...v0.46.0-rc.92) (2023-10-14)
+
+
+### Bug Fixes
+
+* **btp:** fixed z-index of the home list item on expanded and popup modes ([#10717](https://github.com/SAP/fundamental-ngx/issues/10717)) ([a2120b3](https://github.com/SAP/fundamental-ngx/commit/a2120b3475dcb306ea92a2e52cae281e960dd467))
+
+
+
+
+
+# [0.46.0-rc.91](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.90...v0.46.0-rc.91) (2023-10-14)
+
+
+### Bug Fixes
+
+* **btp:** vertical navigation ui bug fixes ([#10716](https://github.com/SAP/fundamental-ngx/issues/10716)) ([bb6323a](https://github.com/SAP/fundamental-ngx/commit/bb6323a78e4d326a6d895ef492d5d91d36004b1e))
+
+
+
+
+
+# [0.46.0-rc.90](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.89...v0.46.0-rc.90) (2023-10-13)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
+# [0.46.0-rc.89](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.88...v0.46.0-rc.89) (2023-10-13)
+
+
+### Features
+
+* **btp:** navigation component implementation ([#10656](https://github.com/SAP/fundamental-ngx/issues/10656)) ([da54e74](https://github.com/SAP/fundamental-ngx/commit/da54e74b969841b221b7b6e193069f79ea2ba787))
+
+
+
+
+
+# [0.46.0-rc.88](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.87...v0.46.0-rc.88) (2023-10-13)
+
+
+### Features
+
+* **btp:** introduced tool header and search field components ([#10706](https://github.com/SAP/fundamental-ngx/issues/10706)) ([abcdb25](https://github.com/SAP/fundamental-ngx/commit/abcdb25341f12af2aac08a1a931133de3f398179))
+
+
+
+
+
 # [0.46.0-rc.87](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.86...v0.46.0-rc.87) (2023-10-11)