+++ libs/core/src/lib/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.46.0-rc.92](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.91...v0.46.0-rc.92) (2023-10-14)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.46.0-rc.91](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.90...v0.46.0-rc.91) (2023-10-14)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.46.0-rc.90](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.89...v0.46.0-rc.90) (2023-10-13)
+
+**Note:** Version bump only for package @fundamental-ngx/core
+
+
+
+
+
+# [0.46.0-rc.89](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.88...v0.46.0-rc.89) (2023-10-13)
+
+
+### Features
+
+* **btp:** navigation component implementation ([#10656](https://github.com/SAP/fundamental-ngx/issues/10656)) ([da54e74](https://github.com/SAP/fundamental-ngx/commit/da54e74b969841b221b7b6e193069f79ea2ba787))
+
+
+
+
+
+# [0.46.0-rc.88](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.87...v0.46.0-rc.88) (2023-10-13)
+
+
+### Features
+
+* **btp:** introduced tool header and search field components ([#10706](https://github.com/SAP/fundamental-ngx/issues/10706)) ([abcdb25](https://github.com/SAP/fundamental-ngx/commit/abcdb25341f12af2aac08a1a931133de3f398179))
+
+
+
+
+
 # [0.46.0-rc.87](https://github.com/SAP/fundamental-ngx/compare/v0.46.0-rc.86...v0.46.0-rc.87) (2023-10-11)
 
 **Note:** Version bump only for package @fundamental-ngx/core